pyproject.toml
[project]
name = "langgraph-quickstart"
version = "0.1.0"
description = "A minimal LangGraph research graph. Clone, set your key, run."
requires-python = ">=3.11,<3.14"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
